feat: refactor welcoming feature into module
